+++ ArkRendering/RenderState.h
@@ -1,5 +1,4 @@
 #pragma once
-
 #include "ArkRendering.h"
 #include "BufferData.h"
 class RenderState
@@ -7,7 +6,7 @@ class RenderState
 public:
 	RenderState(ArkRendering::MaterialInfo const * material, BufferSet const * bufferSet) : mBufferSet(bufferSet), m_material(material) {}
 
-	size_t Size() const { return mBufferSet->Size(); }
+	unsigned int Size() const { return mBufferSet->Size(); }
 	void BindBuffersForDrawing() const { mBufferSet->BindBuffersForDrawing(); }
 	void DisableBuffers() const { mBufferSet->DisableBuffers(); }
